requirements.txt
python-dotenv
beautifulsoup4
langchain-community
langchain-text-splitters
langchain-chroma
langchain-huggingface
langchain-groq
streamlit
streamlit-chat
fabric
